moved engine-loader initializer to app
